refactor: correct file folder and Type names
Refactored the file MVVM folder in model to file (from images), to reflect the inner classes. Also fixed errors due to change Note.kt Type enum NORMAL_TEXT to TEXT, done for better clarity (it is markdown, not .txt)
