+++ app/src/main/java/com/github/onlynotesswent/model/file/FileRepository.kt
@@ -1,4 +1,4 @@
-package com.github.onlynotesswent.model.images
+package com.github.onlynotesswent.model.file
 
 import android.net.Uri
 import com.github.onlynotesswent.model.note.Type
